backend/requirements.txt
affine==2.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
Cartopy==0.24.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contextily==1.6.2
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
Django==5.1.7
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
fonttools==4.57.0
geographiclib==2.0
geopandas==1.0.1
geoplot==0.5.1
geopy==2.4.1
greenlet==3.2.4
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.1
Mako==1.3.10
mapclassify==2.8.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mercantile==1.2.1
msgpack==1.1.0
narwhals==1.35.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.79.0
orjson==3.11.3
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
pyogrio==0.10.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pyxel==2.3.14
PyYAML==6.0.2
pyzmq==26.4.0
rasterio==1.4.3
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentry-sdk==2.38.0
service-identity==24.2.0
setuptools==76.0.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.11.0
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xyzservices==2025.1.0
zope.interface==7.2
